+++ src/components/board/loadmorebutton.js
@@ -1,0 +1,7 @@
+const getLoadMoreButton = function () {
+  return `
+    <button class="load-more" type="button">load more</button>
+  `;
+};
+
+export {getLoadMoreButton};
