+++ src/components/render.js
@@ -1,0 +1,5 @@
+const renderComponent = function (container, html) {
+  container.insertAdjacentHTML(`beforeend`, html);
+};
+
+export default renderComponent;
