+++ webpack.config.js
@@ -10,8 +10,9 @@ module.exports = {
   devtool: `source-map`,
   devServer: {
     contentBase: path.join(__dirname, `public`),
-    publicPath: `http:!//localhost:8080/`,
+    publicPath: `http://localhost:8080/`,
     compress: true,
+    open: true,
     watchContentBase: true
   }
 };
